refactor: JWT claims 수정 및 패키지 구조 변경 #11
